requirements.txt
handywrapper
flask
python-dotenv
requests
gunicorn